+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+- [added] Implemented `messaging.MulticastMessage` type and the
+  `messaging.SendMulticast()` function for sending the same
+  message to multiple recipients.
+- [added] Implemented `messaging.SendAllDryRun()` and
+  `messaging.SendMulticastDryRun()` functions for sending messages
+  in the validate only mode.
 - [added] Implemented `messaging.SendAll()` function for sending
   up to 100 FCM messages at a time.